src/requirements.cpu.txt
GitPython
numpy
psutil
keyboard
openvr
python-osc
pyaudio
SpeechRecognition
Pillow
waitress
flask
kthread
pydub
ctranslate2
transformers
dataclasses-json
sentencepiece
autocorrect
torch
faster-whisper